Rename pipeline.py to indexing.py
'pipeline' was too generic. 'indexing' is the standard IR/RAG term for
the offline flow that turns raw documents into a queryable index
(parse -> TOC -> chunk -> summarize -> embed -> store). Pairs
symmetrically with retrieval.py.
